Disconnect JitConnector from the interpreter initialization logic
